refactor(push-web): split legacy migration parsers
